+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Upcoming Version (WIP)
 
 Improvements:
 * Stabilized consul registration and health checks
+* [MW-1476](https://openlmis.atlassian.net/browse/MW-1476): Add user lockout API and unlock endpoint
 
 Bug fixes:
 * [OLMIS-8223](https://openlmis.atlassian.net/browse/OLMIS-8223): Fix placeholder bug in messages